Inline the closest_stns_within_threshold function
We want to make the database easier to set up and remove dependencies
on things that are not in the ORM. The above function is only used
once in the code, so let's simply inline it. As a side-effect, it's
now more robust, because it doesn't rely on pasting strings together
as parameters and executing text as a query *inside* the function.