Rewrites download iterator for moti and ec
We want to be able to have a download stream optionally get redirected
to a file, but simultaneously maintain the stream for the processing
pipeline. This change ensures that the return value of moti.download
and ec.download are *iterators* and adds a conversion function to
convert them to a readable file-like object. The file-like object can
either be passed to an XML parser, or simply read and written to a
cache file.