Refactors download functions to explicit args
All of the various download functions would only accept a namespace
argument named "args". It was up to the user to actually read through
the code to figure out what arguments need to be supplied. This patch
makes all of the argument lists explicit and modifies the calling
scripts to pass them through.