Makes tz handling consistent across normalize mods
Prior to the introduction of the swob_ml module, there were some
misunderstandings about how to construct timezone aware datetime
objects. Unfortnately, all of the modules incorrectly used
datetime.replace() to add the timezone information. datetime.replace()
does a weird thing where it creates the UTC offset as "Local Mean
Time" which, for Vancouver, is something like 8 hours and 12
minutes... not what we want.

This patch ensures that all normalize() time construction either uses
UTC from the start (if provided), or uses pytz.localize() to add the
timezone information without alterning the time.

Further, this patch ensures that all time values in the normalize()
phase are "converted" (tz info is reassigned and the time value is
adjusted) to UTC before being passed along to the align phase. It's
generally advisable to work in UTC internally to any system.

Refer to here: https://stackoverflow.com/q/1379740/ and here
https://stackoverflow.com/q/35462876/