Update normalize.py
This is a proposed, but untested patch that enables the wamr data stream to hand of certain stations to the correct network where those stations belong to MetroVan. A lookup table is used to change the station_id while the matching logic for the station_id is used to set the network_name to one of MVan or ENV-AQN. These values get passed to the yielded Row object.