Parametrizes caching decorator saving object attrs
"Caching" ORM objects is not really possible. The object is tied to a
database session object, but when the database session is closed, the
object becomed "detached" and no longer has an information available
from it. This patch mitigates that by, rather than caching an ORM
object from the session, one parameterizes the decorator with the
object attributes that should be cached. These attributes are cached
(in a SimpleNamespace object) and can then be reused.