Address several autoflush warnings in SQLA
SQLAlchemy is issuing warnings for querying objects and then using
them (in a read-only context) when the object and related objects are
not added to the session. Since these are all happening in database
code that is inteded to be read-only, we can safely avoid these by
using a no-autoflush session context manager.